+++ includes/ft_striteri.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char *))
+#include "../libft.h"
+
+void	ft_striteri(char *s, void (*f)(size_t, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
