+++ includes/ft_strncmp.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *string1, const char *string2, unsigned int num)
+#include "../libft.h"
+
+int	ft_strncmp(const char *string1, const char *string2, size_t num)
 {
     while (*string1 && *string2 && num--)
     {
